Rename to `pwlogger` and use `ErrAttr`
