+++ browser_context.go
@@ -10,6 +10,7 @@ import (
 	"strings"
 	"sync"
 
+	"github.com/playwright-community/playwright-go/internal/pwlogger"
 	"github.com/playwright-community/playwright-go/internal/safe"
 )
 
@@ -589,7 +590,7 @@ func (b *browserContextImpl) onRoute(route *routeImpl) {
 				return nil, err
 			}, true)
 			if err != nil {
-				logger.Error("could not update interception patterns: %v", err)
+				logger.Error("Could not update interception patterns", pwlogger.ErrAttr(err))
 			}
 		}
 	}
@@ -742,7 +743,7 @@ func (b *browserContextImpl) onWebSocketRoute(wr WebSocketRoute) {
 		b.Unlock()
 		_, err := wr.ConnectToServer()
 		if err != nil {
-			logger.Println(err)
+			logger.Error("Could not connect to WebSocket server", pwlogger.ErrAttr(err))
 		}
 		return
 	}
