+++ channel.go
@@ -3,6 +3,8 @@ package playwright
 import (
 	"encoding/json"
 	"fmt"
+
+	"github.com/playwright-community/playwright-go/internal/pwlogger"
 )
 
 type channel struct {
@@ -77,7 +79,7 @@ func (c *channel) innerSendNoReply(method string, isInternal bool, options ...in
 	}, isInternal)
 	if err != nil {
 		// ignore error actively, log only for debug
-		logger.Error("SendNoReply failed: %v", err)
+		logger.Error("SendNoReply failed", pwlogger.ErrAttr(err))
 	}
 }
