+++ websocket.go
@@ -3,6 +3,8 @@ package playwright
 import (
 	"encoding/base64"
 	"errors"
+
+	"github.com/playwright-community/playwright-go/internal/pwlogger"
 )
 
 type webSocketImpl struct {
@@ -50,7 +52,7 @@ func (ws *webSocketImpl) onFrameSent(opcode float64, data string) {
 	if opcode == 2 {
 		payload, err := base64.StdEncoding.DecodeString(data)
 		if err != nil {
-			logger.Error("could not decode WebSocket.onFrameSent payload: %v", err)
+			logger.Error("could not decode WebSocket.onFrameSent payload", pwlogger.ErrAttr(err))
 			return
 		}
 		ws.Emit("framesent", payload)
@@ -63,7 +65,7 @@ func (ws *webSocketImpl) onFrameReceived(opcode float64, data string) {
 	if opcode == 2 {
 		payload, err := base64.StdEncoding.DecodeString(data)
 		if err != nil {
-			logger.Error("could not decode WebSocket.onFrameReceived payload: %v", err)
+			logger.Error("could not decode WebSocket.onFrameReceived payload", pwlogger.ErrAttr(err))
 			return
 		}
 		ws.Emit("framereceived", payload)
